+++ bgameb/game.py
@@ -2,12 +2,10 @@
 """
 from typing import Optional
 from dataclasses import dataclass, field
+from dataclasses_json import config
 from bgameb.base import Base, log_enable
-from bgameb.types import (
-    COMPONENTS, NONSTUFF, COMPONENTS_TYPES, NONSTUFF_TYPES,
-    STUFF_TYPES, STUFF
-)
-from bgameb.tools import Rules, Turn
+from bgameb.types import COMPONENTS, NONSTUFF, COMPONENTS_TYPES, STUFF
+from bgameb.tools import Rules, Turns
 from bgameb.errors import ComponentClassError
 
 
@@ -16,59 +14,26 @@ class BaseGame(Base):
     """Base class for game
     """
     game_rules: Rules = field(init=False)
-    turn_order: Turn = field(init=False)
+    turn_order: Turns = field(init=False)
 
     def __post_init__(self) -> None:
         super().__post_init__()
         self.game_rules = Rules('game_rules')
-        self.turn_order = Turn('turn_order')
+        self.turn_order = Turns('turn_order')
 
 
 @dataclass
 class Game(BaseGame):
     """Create the game object
     """
+    type_: str = field(
+        default='game',
+        metadata=config(exclude=lambda x: True),  # type: ignore
+        repr=False
+        )
 
     def __post_init__(self) -> None:
         super().__post_init__()
-
-    # def add(
-    #     self,
-    #     component: COMPONENTS_TYPES,
-    #     name: str,
-    #     **kwargs
-    #         ) -> None:
-    #     """Add stuff to game tool
-
-    #     Args:
-    #         name (str): name of added stuff.
-    #         to (str): name of tool, where added
-    #         **kwargs (Dict[str,Any]): dict of named args
-    #     """
-    #     if component in COMPONENTS.keys():
-    #         self._add(COMPONENTS[component], name=name, **kwargs)
-    #         self.logger.info(f'{name} is added to game.')
-    #     else:
-    #         raise ComponentClassError(component, self.logger)
-
-    # def add_to(
-    #     self,
-    #     to: str,
-    #     name: str,
-    #     **kwargs
-    #         ) -> None:
-    #     """Add object to game
-
-    #     Args:
-    #         component (COMPONENTS_TYPES): type of added component.
-    #         name (str): name of added component.
-    #         **kwargs (Dict[str,Any]): dict of named args
-    #     """
-    #     if name in self.keys() and to in self.keys():
-    #         self[to]._increase(name=name, game=self, **kwargs)
-    #         self.logger.info(f'{name} is added to {to}.')
-    #     else:
-    #         raise ComponentClassError(name, self.logger)
 
     def new(
         self,
@@ -77,12 +42,12 @@ class Game(BaseGame):
         target: Optional[str] = None,
         **kwargs
             ) -> None:
-        """Add new component to game object or any game-hosted object
+        """Add new component to game or any stuff to game-hosted tool or player
 
         Args:
             name (str): name of new component (must be unic)
             ctype (COMPONENTS_TYPES): type of new component
-            target (Optional[str], optional): name of class, where placed copy.
+            target (str, optional): name of class, where placed component.
                                               Defaults to None.
 
         Raises:
@@ -92,7 +57,9 @@ class Game(BaseGame):
             if not target:
                 self._add(COMPONENTS[ctype], name=name, **kwargs)
                 self.logger.info(f'{name} is added to game.')
-            elif target in self.keys():
+            elif target in self.keys() \
+                    and self[target].type_ in NONSTUFF.keys() \
+                    and ctype in STUFF.keys():
                 self[target]._add(COMPONENTS[ctype], name=name, **kwargs)
                 self.logger.info(f'{name} is added to {target}.')
             else:
@@ -110,25 +77,27 @@ class Game(BaseGame):
         """Copy stuff from game object to any game-hosted tool or player
 
         Args:
-            source (str): name of copied class
-            target (str): name of class, where placed copy
+            source (str): name of copied stuff
+            target (str): name of tool or player, where placed copy
 
         Raises:
             ComponentClassError: target or source not exist
 
         As result this operation we need two classes inside Game object:
-        - target Component-like class
-        - source Component-like class
+        - target Component-like stuff
+        - source Component-like tool
 
         When method was called, is created new Component with all
         attributes of source class and it placed in target class
         """
-        if source in self.keys() and target in self.keys():
-            self[target]._increase(name=source, game=self, **kwargs)
+        if source in self.keys() \
+                and self[source].type_ in STUFF.keys() \
+                and target in self.keys() \
+                and self[target].type_ in NONSTUFF.keys():
+            self[target].update(name=source, game=self, **kwargs)
             self.logger.info(f'{source} is added to {target}.')
         else:
             raise ComponentClassError((source, target), self.logger)
-
 
 
 if __name__ == '__main__':
@@ -146,6 +115,9 @@ if __name__ == '__main__':
     game.copy('this_rule', 'turn_order')
     game.cards_deck.deal()
     game.turn_order.deal()
+    game.new('blue_shaker', ctype='shaker')
+    game.new('eight', ctype='dice', target='blue_shaker', sides=8, count=10)
+    result = game.blue_shaker.eight.roll()
 
     print(game)
     print('='*20)