+++ bgameb/stuff.py
@@ -26,6 +26,11 @@ class Rule(BaseStuff):
         repr=False
         )
     text: Optional[str] = None
+    type_: str = field(
+        default='rule',
+        metadata=config(exclude=lambda x: True),  # type: ignore
+        repr=False
+        )
 
     def __post_init__(self) -> None:
         super().__post_init__()
@@ -57,6 +62,11 @@ class Dice(BaseStuff):
         repr=False
         )
     sides: int = 2
+    type_: str = field(
+        default='dice',
+        metadata=config(exclude=lambda x: True),  # type: ignore
+        repr=False
+        )
 
     def __post_init__(self) -> None:
         super().__post_init__()
@@ -98,6 +108,11 @@ class Card(BaseStuff):
     tapped: bool = False
     side: Optional[str] = None
     counter: Counter = field(default_factory=Counter)
+    type_: str = field(
+        default='card',
+        metadata=config(exclude=lambda x: True),  # type: ignore
+        repr=False
+        )
 
     def __post_init__(self) -> None:
         super().__post_init__()
@@ -152,8 +167,8 @@ class Card(BaseStuff):
 
 
 STUFF = {
-    'rule': Rule,
-    'dice': Dice,
-    'card': Card,
+    Rule.type_: Rule,
+    Dice.type_: Dice,
+    Card.type_: Card,
     }
 STUFF_TYPES = Literal['rule', 'dice', 'card']