+++ tests/test_game.py
@@ -1,4 +1,5 @@
-import json, pytest
+import json
+import pytest
 from bgameb.game import Game
 from bgameb.base import Components
 from bgameb.types import COMPONENTS
@@ -15,8 +16,10 @@ class TestGame:
         obj_ = Game(name='this_game')
         assert obj_.name == 'this_game', 'not set name for instance'
         assert obj_.is_active, 'wrong is_active'
-        assert isinstance(obj_.turn_order, COMPONENTS['turn']), 'turn_order isnt component'
-        assert isinstance(obj_.game_rules, COMPONENTS['rules']), 'rules isnt component'
+        assert isinstance(obj_.turn_order, COMPONENTS['turns']), \
+            'turn_order isnt component'
+        assert isinstance(obj_.game_rules, COMPONENTS['rules']), \
+            'rules isnt component'
         assert isinstance(obj_, Components), 'isnt component'
 
     def test_game_class_is_converted_to_json(self) -> None:
@@ -25,17 +28,6 @@ class TestGame:
         obj_ = Game(name='game')
         j = json.loads(obj_.to_json())
         assert j['name'] == 'game', 'not converted to json'
-
-    # def test_add_component(self) -> None:
-    #     """Test add tool or stuff to game
-    #     """
-    #     obj_ = Game(name='game')
-    #     for n in COMPONENTS.keys():
-    #         if n == 'rule':
-    #             obj_.add(n, name=n, text='this')
-    #         else:
-    #             obj_.add(n, name=n)
-    #         assert obj_[n].name == n, 'component not added'
 
     def test_add_new_component_to_game(self) -> None:
         """Test new() tool or stuff add to game
@@ -48,29 +40,12 @@ class TestGame:
                 obj_.new(name=n, ctype=n)
             assert obj_[n].name == n, 'component not added'
 
-    # def test_add_component_with_kwargs(self) -> None:
-    #     """Test add with kwargs
-    #     """
-    #     obj_ = Game(name='game')
-    #     obj_.add('dice', name='dice', sides=42)
-    #     assert obj_.dice.sides == 42, 'wrong count'
-
     def test_add_new_component_with_kwargs(self) -> None:
         """Test new() add to game with kwargs
         """
         obj_ = Game(name='game')
         obj_.new('dice', ctype='dice', sides=42)
         assert obj_.dice.sides == 42, 'wrong count'
-
-    # def test_add_new_wrong_component_to_game(self) -> None:
-    #     """Test we cant add notexisted tool or stuff to game
-    #     """
-    #     obj_ = Game(name='game')
-    #     with pytest.raises(
-    #         ComponentClassError,
-    #         match='not a component'
-    #         ):
-    #         obj_.add('chocho', name='this')
 
     def test_add_new_wrong_component_to_game(self) -> None:
         """Test we cant add new notexisted tool or stuff to game
@@ -79,17 +54,8 @@ class TestGame:
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
+                ):
             obj_.new('chocho', ctype='this')
-
-    # def test_add_to(self) -> None:
-    #     """Test add_to() method
-    #     """
-    #     obj_ = Game(name='game')
-    #     obj_.add('dice', name='that')
-    #     obj_.add('shaker', name='this')
-    #     obj_.add_to('this', 'that', count=10)
-    #     assert obj_.this.that.count == 10, 'not added'
 
     def test_add_new_to_existed_object(self) -> None:
         """Test new() method for add new object to existed component
@@ -108,7 +74,7 @@ class TestGame:
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
+                ):
             obj_.new('that', ctype='what', target='this', count=10)
 
     def test_add_new_wrong_component_to_notexisted_object(self) -> None:
@@ -118,7 +84,7 @@ class TestGame:
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
+                ):
             obj_.new('that', ctype='what', target='this', count=10)
 
     def test_add_new_component_to_notexisted_object(self) -> None:
@@ -128,25 +94,30 @@ class TestGame:
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
+                ):
             obj_.new('that', ctype='dice', target='this', count=10)
 
-    # def test_add_notexisted(self) -> None:
-    #     """Test add_to() notexisted stuff or tool
-    #     """
-    #     obj_ = Game(name='game')
-    #     obj_.add('dice', name='that')
-    #     obj_.add('shaker', name='this')
-    #     with pytest.raises(
-    #         ComponentClassError,
-    #         match='not a component'
-    #         ):
-    #         obj_.add_to('this', 'what', count=10)
-    #     with pytest.raises(
-    #         ComponentClassError,
-    #         match='not a component'
-    #         ):
-    #         obj_.add_to('what', 'that', count=10)
+    def test_add_new_stuff_to_nontools(self) -> None:
+        """Test new() cant add new stuff to nontools/players
+        """
+        obj_ = Game(name='game')
+        obj_.new('this', ctype='dice')
+        with pytest.raises(
+            ComponentClassError,
+            match='not a component'
+                ):
+            obj_.new('that', ctype='dice', target='this')
+
+    def test_add_new_nonstuff_to_tools(self) -> None:
+        """Test new() cant add new stuff to nontools/players
+        """
+        obj_ = Game(name='game')
+        obj_.new('this', ctype='shaker')
+        with pytest.raises(
+            ComponentClassError,
+            match='not a component'
+                ):
+            obj_.new('that', ctype='shaker', target='this', count=10)
 
     def test_copy_component_from_game_to_another_component(self) -> None:
         """Test copy() existed component to another component
@@ -176,7 +147,7 @@ class TestGame:
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
+                ):
             obj_.copy('that', 'this', count=6)
 
     def test_copy_component_to_notexisted_tool(self) -> None:
@@ -187,7 +158,7 @@ class TestGame:
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
+                ):
             obj_.copy('that', 'this', count=6)
 
     def test_add_rule_to_rules(self) -> None:
@@ -196,24 +167,29 @@ class TestGame:
         obj_ = Game(name='game')
         obj_.new('this', ctype='rule', text='that')
         obj_.copy('this', 'game_rules')
-        assert isinstance(obj_.game_rules.this, COMPONENTS['rule']), 'wrong type of rule'
+        assert isinstance(obj_.game_rules.this, COMPONENTS['rule']), \
+            'wrong type of rule'
         assert obj_.game_rules.this.name == 'this', 'wrong name of rule'
         assert obj_.game_rules.this.text == 'that', 'wrong text rule'
 
-    @pytest.mark.skip('# TODO: define logic')
-    def test_add_not_stuff_or_not_to_tool(self) -> None:
-        """Test add_to() missed type
+    def test_copy_not_stuff_or_not_to_tool(self) -> None:
+        """Test copy() missed type
         """
         obj_ = Game(name='game')
-        obj_.add('dice', name='that')
-        obj_.add('shaker', name='this')
+        obj_.new('that', 'shaker')
+        obj_.new('this', 'dice')
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
-            obj_.add_to('this', 'this', count=10)
+                ):
+            obj_.copy('this', 'this')
         with pytest.raises(
             ComponentClassError,
             match='not a component'
-            ):
-            obj_.add_to('that', 'that', count=10)
+                ):
+            obj_.copy('that', 'that', count=10)
+        with pytest.raises(
+            ComponentClassError,
+            match='not a component'
+                ):
+            obj_.copy('that', 'this', count=10)